requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 图像与数据处理
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# 数据增强
albumentations>=1.3.0

# 预训练模型
timm>=0.9.0

# 交互页面
streamlit>=1.28.0

# 配置文件
pyyaml>=6.0

# 进度条
tqdm>=4.66.0

# 绘图
matplotlib>=3.7.0